Refactor CeleryPayload as an inner class in CeleryProcessingService
